backend/routers/video.py
from fastapi import APIRouter, UploadFile, File, HTTPException
from pydantic import BaseModel
import uuid
import os

router = APIRouter(prefix="/video", tags=["video"])


@router.post("/upload")
async def upload_video(file: UploadFile = File(...)):
    # Placeholder for upload logic
    return {"filename": file.filename}


@router.post("/analyze")
async def analyze_video(video_id: str):
    # Placeholder for analysis logic
    return {"status": "processing", "video_id": video_id}
