backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.122.0",
    "google-cloud-storage>=3.6.0",
    "google-genai>=1.52.0",
    "numpy>=2.3.5",
    "opencv-python-headless>=4.11.0.86",
    "pillow>=12.0.0",
    "python-multipart>=0.0.20",
    "uvicorn>=0.38.0",
]
